# Add information to `barman show-backup` about incremental backups