--- a/barman/server.py
+++ b/barman/server.py
@@ -3037,6 +3037,9 @@ class Server(RemoteStatusMixin):
          * the Server.get_wal_info() return value
          * the context in the catalog (if available)
          * the retention policy status
+         * the copy statistics
+         * the incremental backups information
+         * extra backup.info properties
 
         :param backup_info: the target backup
         :rtype dict: all information about a backup
@@ -3050,27 +3053,24 @@ class Server(RemoteStatusMixin):
                 next_backup = self.backup_manager.get_next_backup(
                     backup_ext_info["backup_id"]
                 )
+                backup_ext_info["previous_backup_id"] = None
+                backup_ext_info["next_backup_id"] = None
                 if previous_backup:
                     backup_ext_info["previous_backup_id"] = previous_backup.backup_id
-                else:
-                    backup_ext_info["previous_backup_id"] = None
                 if next_backup:
                     backup_ext_info["next_backup_id"] = next_backup.backup_id
-                else:
-                    backup_ext_info["next_backup_id"] = None
             except UnknownBackupIdException:
                 # no next_backup_id and previous_backup_id items
                 # means "Not available"
                 pass
             backup_ext_info.update(self.get_wal_info(backup_info))
+
+            backup_ext_info["retention_policy_status"] = None
             if self.enforce_retention_policies:
                 policy = self.config.retention_policy
                 backup_ext_info["retention_policy_status"] = policy.backup_status(
                     backup_info.backup_id
                 )
-            else:
-                backup_ext_info["retention_policy_status"] = None
-
             # Check any child timeline exists
             children_timelines = self.get_children_timelines(
                 backup_ext_info["timeline"], forked_after=backup_info.end_xlog
@@ -3078,6 +3078,43 @@ class Server(RemoteStatusMixin):
 
             backup_ext_info["children_timelines"] = children_timelines
 
+            # If copy statistics are available
+            copy_stats = backup_ext_info.get("copy_stats")
+            if copy_stats:
+                analysis_time = copy_stats.get("analysis_time", 0)
+                if analysis_time >= 1:
+                    backup_ext_info["analysis_time"] = analysis_time
+                copy_time = copy_stats.get("copy_time", 0)
+                if copy_time > 0:
+                    backup_ext_info["copy_time"] = copy_time
+                    dedup_size = backup_ext_info.get("deduplicated_size", 0)
+                    if dedup_size > 0:
+                        estimated_throughput = dedup_size / copy_time
+                        backup_ext_info["estimated_throughput"] = estimated_throughput
+                        number_of_workers = copy_stats.get("number_of_workers", 1)
+                        if number_of_workers > 1:
+                            backup_ext_info["number_of_workers"] = number_of_workers
+
+            backup_chain = [backup for backup in backup_info.walk_to_root()]
+            chain_size = len(backup_chain)
+            # last is root
+            root_backup_info = backup_chain[-1]
+            # "Incremental" backups
+            backup_ext_info["root_backup_id"] = root_backup_info.backup_id
+            backup_ext_info["chain_size"] = chain_size
+            # Properties added to the result dictionary
+            backup_ext_info["backup_type"] = backup_info.backup_type
+            backup_ext_info["deduplication_ratio"] = backup_info.deduplication_ratio
+            # A new field "cluster_size" was added to backup.info to be
+            # able to calculate the resource saved by "incremental" backups
+            # introduced in Postgres 17.
+            # To keep backward compatibility between versions, barman relies
+            # on two possible values to calculate "est_dedup_size",
+            # "size" being used for older versions when "cluster_size"
+            # is non existent (None).
+            backup_ext_info["est_dedup_size"] = (
+                backup_ext_info["cluster_size"] or backup_ext_info["size"]
+            ) * backup_ext_info["deduplication_ratio"]
         return backup_ext_info
 
     def show_backup(self, backup_info):