… barman/output.py
… tests/test_output.py