Now that Postgres 17 have the ability to take incremental backups with pg_basebackup, there is a need to add new information to the show-backup command. 

This PR will:
1. Add fields to the `get_backup_ext_info` method from `Server` so we have extra information to be reused in the outputs.
2. Modify `output.py` to accommodate the new information that came with the introduction of incremental backups.
3. Fix tests broken by this implementation
4. Add unit tests according to use cases.


Note: 
The code in the output.py have to deal with two inputs: the result of the `get_backup_ext_info` method call **(which is a dictionary with the extra fields mentioned in 1.)** or backup_info.to_dict() **(which is for cloud-show-backup)**. 
The latter will not have the new fields mentioned in 1.  This is also the case for `--format json` for barman-cloud-backup-show. It relies on the backup_info.to_json() method. The code implemented is a little repetitive so we do not break the barman-cloud-backup-show command.

References: BAR-208 & BAR-214