@barthisrael 2024-08-01
I have performed some tests to check how things look like.

I have these backups, both from Postgres 17. The custom names of the backup should give you an idea of what they are 😄 :

```console
$ barman list-backup all
pg17 20240801T201448 'pg_incremental' - I - Thu Aug  1 18:14:53 2024 - Size: 365.9 MiB - WAL Size: 0 B - WAITING_FOR_WALS
pg17 20240801T201244 'pg_full' - F - Thu Aug  1 18:12:47 2024 - Size: 251.5 MiB - WAL Size: 656.0 MiB
support 20240801T201405 'rsync_incremental' - R - Thu Aug  1 18:14:14 2024 - Size: 581.3 MiB - WAL Size: 32.0 MiB
support 20240801T201228 'rsync_full' - R - Thu Aug  1 18:12:36 2024 - Size: 235.2 MiB - WAL Size: 672.0 MiB
```

This is how `show-backup` of a rsync full looks like:

```console
$ barman show-backup support rsync_full
Backup 20240801T201228:
  Backup Name            : rsync_full
  Server Name            : support
  System Id              : 7382227563826631291
  Status                 : DONE
  PostgreSQL Version     : 170000
  PGDATA directory       : /var/lib/pgsql/17/data
  Estimated Cluster Size : 219.0 MiB

  Server information:
    Checksums            : off
    WAL summarizer       : on

  Tablespaces:
    custom_ts_1          : /var/lib/pgsql/17/tablespaces/custom_ts_1 (oid: 24678)
    custom_ts_2          : /var/lib/pgsql/17/tablespaces/custom_ts_2 (oid: 24686)

  Base backup information:
    Backup Method        : rsync-concurrent
    Backup Size          : 219.2 MiB (235.2 MiB with WALs)
    WAL Size             : 16.0 MiB
    Resource savings     : 0 B (0.00%)
    Timeline             : 1
    Begin WAL            : 000000010000000600000052
    End WAL              : 000000010000000600000052
    WAL number           : 1
    Begin time           : 2024-08-01 18:12:28.516824+00:00
    End time             : 2024-08-01 18:12:36.132905+00:00
    Copy time            : 5 seconds + 1 second startup
    Estimated throughput : 38.6 MiB/s
    Begin Offset         : 40
    End Offset           : 344
    Begin LSN            : 6/52000028
    End LSN              : 6/52000158

  WAL information:
    No of files          : 42
    Disk usage           : 672.0 MiB
    WAL rate             : 1573.45/hour
    Last available       : 00000001000000060000007C

  Catalog information:
    Retention Policy     : not enforced
    Previous Backup      : - (this is the oldest base backup)
    Next Backup          : 20240801T201405
```

This is the diff compared to `master` branch:

```diff
8c8,12
<   Checksums              : off
---
>   Estimated Cluster Size : 219.0 MiB
>
>   Server information:
>     Checksums            : off
>     WAL summarizer       : on
15,16c19,22
<     Disk usage           : 219.2 MiB (235.2 MiB with WALs)
<     Incremental size     : 219.2 MiB (-0.00%)
---
>     Backup Method        : rsync-concurrent
>     Backup Size          : 219.2 MiB (235.2 MiB with WALs)
>     WAL Size             : 16.0 MiB
>     Resource savings     : 0 B (0.00%)
```

This is how rsync incremental looks like:

```console
$ barman show-backup support rsync_incremental
Backup 20240801T201405:
  Backup Name            : rsync_incremental
  Server Name            : support
  System Id              : 7382227563826631291
  Status                 : DONE
  PostgreSQL Version     : 170000
  PGDATA directory       : /var/lib/pgsql/17/data
  Estimated Cluster Size : 565.1 MiB

  Server information:
    Checksums            : off
    WAL summarizer       : on

  Tablespaces:
    custom_ts_1          : /var/lib/pgsql/17/tablespaces/custom_ts_1 (oid: 24678)
    custom_ts_2          : /var/lib/pgsql/17/tablespaces/custom_ts_2 (oid: 24686)

  Base backup information:
    Backup Method        : rsync-concurrent
    Backup Size          : 347.5 MiB (363.5 MiB with WALs)
    WAL Size             : 16.0 MiB
    Resource savings     : 217.7 MiB (38.53%)
    Timeline             : 1
    Begin WAL            : 00000001000000060000007C
    End WAL              : 00000001000000060000007C
    WAL number           : 1
    Begin time           : 2024-08-01 18:14:05.743646+00:00
    End time             : 2024-08-01 18:14:14.509817+00:00
    Copy time            : 6 seconds + 1 second startup
    Estimated throughput : 57.8 MiB/s
    Begin Offset         : 37752
    End Offset           : 1029312
    Begin LSN            : 6/7C009378
    End LSN              : 6/7C0FB4C0

  WAL information:
    No of files          : 2
    Disk usage           : 32.0 MiB
    WAL rate             : 289.85/hour
    Last available       : 00000001000000060000007E

  Catalog information:
    Retention Policy     : not enforced
    Previous Backup      : 20240801T201228
    Next Backup          : - (this is the latest base backup)
```

This is the diff compared to `master` branch:

```diff
8c8,12
<   Checksums              : off
---
>   Estimated Cluster Size : 565.1 MiB
>
>   Server information:
>     Checksums            : off
>     WAL summarizer       : on
15,16c19,22
<     Disk usage           : 565.3 MiB (581.3 MiB with WALs)
<     Incremental size     : 347.5 MiB (-38.53%)
---
>     Backup Method        : rsync-concurrent
>     Backup Size          : 347.5 MiB (363.5 MiB with WALs)
>     WAL Size             : 16.0 MiB
>     Resource savings     : 217.7 MiB (38.53%)
```

This is how pg full looks like:

```console
$ barman show-backup pg17 pg_full
Backup 20240801T201244:
  Backup Name            : pg_full
  Server Name            : pg17
  System Id              : 7382227563826631291
  Status                 : DONE
  PostgreSQL Version     : 170000
  PGDATA directory       : /var/lib/pgsql/17/data
  Estimated Cluster Size : 219.0 MiB

  Server information:
    Checksums            : off
    WAL summarizer       : on

  Tablespaces:
    custom_ts_1          : /var/lib/pgsql/17/tablespaces/custom_ts_1 (oid: 24678)
    custom_ts_2          : /var/lib/pgsql/17/tablespaces/custom_ts_2 (oid: 24686)

  Base backup information:
    Backup Method        : postgres
    Backup Type          : full
    Children Backup Id(s): ['20240801T201448']
    Backup Size          : 219.5 MiB (251.5 MiB with WALs)
    WAL Size             : 32.0 MiB
    Timeline             : 1
    Begin WAL            : 000000010000000600000054
    End WAL              : 000000010000000600000055
    WAL number           : 2
    Begin time           : 2024-08-01 18:12:44.372221+00:00
    End time             : 2024-08-01 18:12:47.489739+00:00
    Copy time            : 3 seconds
    Estimated throughput : 70.8 MiB/s
    Begin Offset         : 40
    End Offset           : 0
    Begin LSN            : 6/54000028
    End LSN              : 6/55000000

  WAL information:
    No of files          : 41
    Disk usage           : 656.0 MiB
    WAL rate             : 1227.44/hour
    Last available       : 00000001000000060000007E

  Catalog information:
    Retention Policy     : VALID
    Previous Backup      : - (this is the oldest base backup)
    Next Backup          : 20240801T201448
```

This is the diff compared to `master` branch:

```diff
8c8,12
<   Checksums              : off
---
>   Estimated Cluster Size : 219.0 MiB
>
>   Server information:
>     Checksums            : off
>     WAL summarizer       : on
15,16c19,23
<     Disk usage           : 219.5 MiB (251.5 MiB with WALs)
<     Incremental size     : 219.5 MiB (-0.00%)
---
>     Backup Method        : postgres
>     Backup Type          : full
>     Children Backup Id(s): ['20240801T201448']
>     Backup Size          : 219.5 MiB (251.5 MiB with WALs)
>     WAL Size             : 32.0 MiB
```

This is how pg incremental looks like:

```console
$ barman show-backup pg17 pg_incremental
Backup 20240801T201448:
  Backup Name            : pg_incremental
  Server Name            : pg17
  System Id              : 7382227563826631291
  Status                 : WAITING_FOR_WALS
  PostgreSQL Version     : 170000
  PGDATA directory       : /var/lib/pgsql/17/data
  Estimated Cluster Size : 565.1 MiB

  Server information:
    Checksums            : off
    WAL summarizer       : on

  Tablespaces:
    custom_ts_1          : /var/lib/pgsql/17/tablespaces/custom_ts_1 (oid: 24678)
    custom_ts_2          : /var/lib/pgsql/17/tablespaces/custom_ts_2 (oid: 24686)

  Base backup information:
    Backup Method        : postgres
    Backup Type          : incremental
    Root Backup Id       : 20240801T201244
    Parent Backup Id     : 20240801T201244
    Backup chain size    : 2
    Backup Size          : 349.9 MiB (365.9 MiB with WALs)
    WAL Size             : 16.0 MiB
    Resource savings     : 215.3 MiB (38.09%)
    Timeline             : 1
    Begin WAL            : 00000001000000060000007E
    End WAL              : 00000001000000060000007F
    WAL number           : 1
    Begin time           : 2024-08-01 18:14:48.599869+00:00
    End time             : 2024-08-01 18:14:53.684562+00:00
    Copy time            : 5 seconds
    Estimated throughput : 69.2 MiB/s
    Begin Offset         : 40
    End Offset           : 0
    Begin LSN            : 6/7E000028
    End LSN              : 6/7F000000

  WAL information:
    No of files          : 0
    Disk usage           : 0 B
    Last available       : 00000001000000060000007E

  Catalog information:
    Retention Policy     : VALID
    Previous Backup      : 20240801T201244
    Next Backup          : - (this is the latest base backup)
```

This is the diff compared to `master` branch`:

```diff
8c8,12
<   Checksums              : off
---
>   Estimated Cluster Size : 565.1 MiB
>
>   Server information:
>     Checksums            : off
>     WAL summarizer       : on
15,16c19,26
<     Disk usage           : 349.9 MiB (365.9 MiB with WALs)
<     Incremental size     : 349.9 MiB (-0.00%)
---
>     Backup Method        : postgres
>     Backup Type          : incremental
>     Root Backup Id       : 20240801T201244
>     Parent Backup Id     : 20240801T201244
>     Backup chain size    : 2
>     Backup Size          : 349.9 MiB (365.9 MiB with WALs)
>     WAL Size             : 16.0 MiB
>     Resource savings     : 215.3 MiB (38.09%)
```