@barthisrael 2024-08-08
I took a backup with `pg_basebackup` using Barman 3.10.1.

* Console output with `3.10.1`:

```console
Backup 20240808T191728:
  Server Name            : pg17
  System Id              : 7382227563826631291
  Status                 : DONE
  PostgreSQL Version     : 170000
  PGDATA directory       : /var/lib/pgsql/17/data

  Base backup information:
    Disk usage           : 25.3 MiB (57.3 MiB with WALs)
    Incremental size     : 25.3 MiB (-0.00%)
    Timeline             : 1
    Begin WAL            : 0000000100000007000000F2
    End WAL              : 0000000100000007000000F3
    WAL number           : 2
    Begin time           : 2024-08-08 17:17:28.190827+00:00
    End time             : 2024-08-08 17:17:33.493005+00:00
    Copy time            : 5 seconds
    Estimated throughput : 4.8 MiB/s
    Begin Offset         : 40
    End Offset           : 0
    Begin LSN            : 7/F2000028
    End LSN              : 7/F3000000

  WAL information:
    No of files          : 22
    Disk usage           : 352.0 MiB
    WAL rate             : 1223.59/hour
    Last available       : 000000010000000800000009

  Catalog information:
    Retention Policy     : VALID
    Previous Backup      : - (this is the oldest base backup)
    Next Backup          : - (this is the latest base backup)
```

* Console output with this dev branch:

```console
Backup 20240808T191728:
  Server Name            : pg17
  System Id              : 7382227563826631291
  Status                 : DONE
  PostgreSQL Version     : 170000
  PGDATA directory       : /var/lib/pgsql/17/data

  Base backup information:
    Backup Method        : postgres
    Backup Type          : full
    Backup Size          : 25.3 MiB (57.3 MiB with WALs)
    WAL Size             : 32.0 MiB
    Timeline             : 1
    Begin WAL            : 0000000100000007000000F2
    End WAL              : 0000000100000007000000F3
    WAL number           : 2
    Begin time           : 2024-08-08 17:17:28.190827+00:00
    End time             : 2024-08-08 17:17:33.493005+00:00
    Copy time            : 5 seconds
    Estimated throughput : 4.8 MiB/s
    Begin Offset         : 40
    End Offset           : 0
    Begin LSN            : 7/F2000028
    End LSN              : 7/F3000000

  WAL information:
    No of files          : 22
    Disk usage           : 352.0 MiB
    WAL rate             : 1223.59/hour
    Last available       : 000000010000000800000009

  Catalog information:
    Retention Policy     : VALID
    Previous Backup      : - (this is the oldest base backup)
    Next Backup          : - (this is the latest base backup)
```

* The diff between them:

```diff
9,10c9,12
<     Disk usage           : 25.3 MiB (57.3 MiB with WALs)
<     Incremental size     : 25.3 MiB (-0.00%)
---
>     Backup Method        : postgres
>     Backup Type          : full
>     Backup Size          : 25.3 MiB (57.3 MiB with WALs)
>     WAL Size             : 32.0 MiB
```

This looks fine.

* JSON output with 3.10.1:

```json
{
  "pg17": {
    "backup_id": "20240808T191728",
    "base_backup_information": {
      "analysis_time": "less than one second",
      "analysis_time_seconds": 0,
      "begin_lsn": "7/F2000028",
      "begin_offset": 40,
      "begin_time": "2024-08-08 17:17:28.190827+00:00",
      "begin_time_timestamp": "1723137448",
      "begin_wal": "0000000100000007000000F2",
      "copy_time": "5 seconds",
      "copy_time_seconds": 5.270819,
      "disk_usage": "25.3 MiB",
      "disk_usage_bytes": 26536648,
      "disk_usage_with_wals": "57.3 MiB",
      "disk_usage_with_wals_bytes": 60091080,
      "end_lsn": "7/F3000000",
      "end_offset": 0,
      "end_time": "2024-08-08 17:17:33.493005+00:00",
      "end_time_timestamp": "1723137453",
      "end_wal": "0000000100000007000000F3",
      "incremental_size": "25.3 MiB",
      "incremental_size_bytes": 26536648,
      "incremental_size_ratio": "-0.00%",
      "number_of_workers": 1,
      "throughput": "4.8 MiB/s",
      "throughput_bytes": 5034634.655449181,
      "timeline": 1
    },
    "catalog_information": {
      "next_backup": "- (this is the latest base backup)",
      "previous_backup": "- (this is the oldest base backup)",
      "retention_policy": "VALID"
    },
    "pgdata_directory": "/var/lib/pgsql/17/data",
    "postgresql_version": 170000,
    "status": "DONE",
    "tablespaces": [],
    "wal_information": {
      "compression_ratio": 0,
      "disk_usage": "352.0 MiB",
      "disk_usage_bytes": 369098752,
      "last_available": "000000010000000800000009",
      "no_of_files": 22,
      "timelines": [],
      "wal_rate": "1223.59/hour",
      "wal_rate_per_second": 0.339887317436287
    }
  }
}
```

* JSON output with this dev branch:

```json
{
  "_EXCEPTION": [
    "float() argument must be a string or a number, not 'NoneType'\nSee log file for more details."
  ],
  "pg17": {
    "backup_id": "20240808T191728",
    "server_information": {
      "data_checksums": null,
      "summarize_wal": null
    },
    "status": "DONE",
    "system_id": "7382227563826631291"
  }
}
```

This breaks because of this:

```python
See log file for more details.
Traceback (most recent call last):
  File "/home/vagrant/Repositories/edb/barman/barman/cli.py", line 2429, in main
    args.func(args)
  File "/home/vagrant/Repositories/edb/barman/barman/cli.py", line 1400, in show_backup
    server.show_backup(backup_info)
  File "/home/vagrant/Repositories/edb/barman/barman/server.py", line 3128, in show_backup
    output.result("show_backup", backup_ext_info)
  File "/home/vagrant/Repositories/edb/barman/barman/output.py", line 329, in result
    _dispatch(_writer, "result", command, *args, **kwargs)
  File "/home/vagrant/Repositories/edb/barman/barman/output.py", line 198, in _dispatch
    return handler(*args, **kwargs)
  File "/home/vagrant/Repositories/edb/barman/barman/output.py", line 1614, in result_show_backup
    cluster_size=pretty_size(data["cluster_size"]),
  File "/home/vagrant/Repositories/edb/barman/barman/utils.py", line 178, in pretty_size
    size = float(size)
TypeError: float() argument must be a string or a number, not 'NoneType'
```

If I work around that just to check how it would look like:

```json
{
  "pg17": {
    "backup_id": "20240808T191728",
    "base_backup_information": {
      "backup_method": "postgres",
      "backup_size": "25.3 MiB",
      "backup_size_bytes": 26536648,
      "backup_size_with_wals": "57.3 MiB",
      "backup_size_with_wals_bytes": 60091080,
      "begin_lsn": "7/F2000028",
      "begin_offset": 40,
      "begin_time": "2024-08-08 17:17:28.190827+00:00",
      "begin_time_timestamp": "1723137448",
      "begin_wal": "0000000100000007000000F2",
      "copy_time": "5 seconds",
      "copy_time_seconds": 5.270819,
      "end_lsn": "7/F3000000",
      "end_offset": 0,
      "end_time": "2024-08-08 17:17:33.493005+00:00",
      "end_time_timestamp": "1723137453",
      "end_wal": "0000000100000007000000F3",
      "throughput": "4.8 MiB/s",
      "throughput_bytes": 5034634,
      "timeline": 1,
      "wal_num": 2,
      "wal_size": "32.0 MiB",
      "wal_size_bytes": 33554432
    },
    "catalog_information": {
      "backup_type": "full",
      "children_backup_ids": null,
      "next_backup": "- (this is the latest base backup)",
      "previous_backup": "- (this is the oldest base backup)",
      "retention_policy": "VALID"
    },
    "pgdata_directory": "/var/lib/pgsql/17/data",
    "postgresql_version": 170000,
    "server_information": {
      "data_checksums": null,
      "summarize_wal": null
    },
    "status": "DONE",
    "system_id": "7382227563826631291",
    "tablespaces": [],
    "wal_information": {
      "compression_ratio": 0,
      "disk_usage": "352.0 MiB",
      "disk_usage_bytes": 369098752,
      "last_available": "000000010000000800000009",
      "no_of_files": 22,
      "timelines": [],
      "wal_rate": "1223.59/hour",
      "wal_rate_per_second": 0.339887317436287
    }
  }
}
```

And the diff would be:

```diff
5,6c5,9
<       "analysis_time": "less than one second",
<       "analysis_time_seconds": 0,
---
>       "backup_method": "postgres",
>       "backup_size": "25.3 MiB",
>       "backup_size_bytes": 26536648,
>       "backup_size_with_wals": "57.3 MiB",
>       "backup_size_with_wals_bytes": 60091080,
14,17d16
<       "disk_usage": "25.3 MiB",
<       "disk_usage_bytes": 26536648,
<       "disk_usage_with_wals": "57.3 MiB",
<       "disk_usage_with_wals_bytes": 60091080,
23,26d21
<       "incremental_size": "25.3 MiB",
<       "incremental_size_bytes": 26536648,
<       "incremental_size_ratio": "-0.00%",
<       "number_of_workers": 1,
28,29c23,27
<       "throughput_bytes": 5034634.655449181,
<       "timeline": 1
---
>       "throughput_bytes": 5034634,
>       "timeline": 1,
>       "wal_num": 2,
>       "wal_size": "32.0 MiB",
>       "wal_size_bytes": 33554432
31a30,31
>       "backup_type": "full",
>       "children_backup_ids": null,
37a40,43
>     "server_information": {
>       "data_checksums": null,
>       "summarize_wal": null
>     },
38a45
>     "system_id": "7382227563826631291",
```

I think this would also look fine?